rugby-analytics/requirements.txt
requests
beautifulsoup4
pandas
SQLAlchemy
psycopg2-binary
python-dotenv
